Minors changes

+++ src/test/java/com/foursquare/fongo/FongoTest.java
@@ -13,7 +13,6 @@ import com.mongodb.DBCursor;
 import com.mongodb.DBObject;
 import com.mongodb.DBRef;
 import com.mongodb.FongoDBCollection;
-import com.mongodb.MongoClient;
 import com.mongodb.MongoException;
 import com.mongodb.QueryBuilder;
 import com.mongodb.WriteConcern;
@@ -58,7 +57,7 @@ public class FongoTest {
     assertNotNull(collection);
     assertSame("getCollection should be idempotent", collection, db.getCollection("coll"));
     assertSame("getCollection should be idempotent", collection, db.getCollectionFromString("coll"));
-    assertEquals(newHashSet("coll"), db.getCollectionNames());
+    assertEquals(newHashSet("coll", "system.indexes", "system.users"), db.getCollectionNames());
   }
 
   @Test
@@ -66,7 +65,7 @@ public class FongoTest {
     Fongo fongo = newFongo();
     DB db = fongo.getDB("db");
     db.createCollection("coll", null);
-    assertEquals(Collections.singleton("coll"), db.getCollectionNames());
+    assertEquals(new HashSet<String>(Arrays.asList("coll", "system.indexes", "system.users")), db.getCollectionNames());
   }
 
   @Test
@@ -396,6 +395,20 @@ public class FongoTest {
         new BasicDBObject("a", 2).append("_id", 5),
         new BasicDBObject("a", 2).append("_id", 4)
     ), cursor.toArray());
+  }
+
+  @Test
+  public void testCompoundSortFindAndModify() {
+    DBCollection collection = newCollection();
+    collection.insert(new BasicDBObject("a", 1).append("_id", 1));
+    collection.insert(new BasicDBObject("a", 2).append("_id", 5));
+    collection.insert(new BasicDBObject("a", 1).append("_id", 2));
+    collection.insert(new BasicDBObject("a", 2).append("_id", 4));
+    collection.insert(new BasicDBObject("a", 1).append("_id", 3));
+
+    DBObject object = collection.findAndModify(null, new BasicDBObject("a", 1).append("_id", -1), new BasicDBObject("date", 1));
+    assertEquals(
+        new BasicDBObject("_id", 3).append("a", 1), object);
   }
 
   @Test
@@ -816,7 +829,7 @@ public class FongoTest {
   }
 
   @Test
-  public void testSortByEmeddedKey() {
+  public void testSortByEmbeddedKey() {
     DBCollection collection = newCollection();
     collection.insert(new BasicDBObject("_id", 1).append("a", new BasicDBObject("b", 1)));
     collection.insert(new BasicDBObject("_id", 2).append("a", new BasicDBObject("b", 2)));
@@ -890,7 +903,7 @@ public class FongoTest {
     DBCollection collection = newCollection();
     collection.insert(new BasicDBObject());
     collection.drop();
-    assertEquals("Collection should have no data", 0, collection.count());
+    assertEquals("Collection should have no data", 0.D, collection.count(), 0.D);
     assertFalse("Collection shouldn't exist in DB", collection.getDB().getCollectionNames().contains(collection.getName()));
   }
 
@@ -900,7 +913,7 @@ public class FongoTest {
     DBCollection collection = fongo.getDB("db").getCollection("coll");
     collection.insert(new BasicDBObject());
     fongo.dropDatabase("db");
-    assertEquals("Collection should have no data", 0, collection.count());
+    assertEquals("Collection should have no data", 0.D, collection.count(), 0.D);
     assertFalse("Collection shouldn't exist in DB", collection.getDB().getCollectionNames().contains(collection.getName()));
     assertFalse("DB shouldn't exist in fongo", fongo.getDatabaseNames().contains("db"));
   }
@@ -925,10 +938,30 @@ public class FongoTest {
     db.getCollection("coll2");
     int dropCount = 0;
     for (String name : db.getCollectionNames()) {
-      db.getCollection(name).drop();
-      dropCount++;
+      if (!name.startsWith("system.")) {
+        db.getCollection(name).drop();
+        dropCount++;
+      }
     }
     assertEquals("should drop two collections", 2, dropCount);
+  }
+
+  @Test
+  public void testDropCollectionsPermitReuseOfDBCollection() throws Exception {
+    DB db = newFongo().getDB("db");
+    int startingCollectionSize = db.getCollectionNames().size();
+    DBCollection coll1 = db.getCollection("coll1");
+    DBCollection coll2 = db.getCollection("coll2");
+    assertEquals(startingCollectionSize + 2, db.getCollectionNames().size());
+
+    // when
+    coll1.drop();
+    coll2.drop();
+    assertEquals(startingCollectionSize + 0, db.getCollectionNames().size());
+
+    // Insert a value must create the collection.
+    coll1.insert(new BasicDBObject("_id", 1));
+    assertEquals(startingCollectionSize + 1, db.getCollectionNames().size());
   }
 
   @Test
@@ -937,25 +970,37 @@ public class FongoTest {
   }
 
   @Test
-  public void testUndefinedCommand() {
+  public void testForceError() throws Exception {
     Fongo fongo = newFongo();
     DB db = fongo.getDB("db");
-    CommandResult result = db.command("undefined");
-    assertEquals("ok should always be defined", false, result.get("ok"));
-    assertEquals("undefined command: { \"undefined\" : true}", result.get("err"));
+    CommandResult result = db.command("forceerror");
+    System.out.print(result);
+    assertEquals("ok should always be defined", 0.0, result.get("ok"));
+    assertEquals("exception: forced error", result.get("errmsg"));
+    assertEquals(10038, result.get("code"));
   }
 
   @Test
-  public void testCountCommand() {
-    DBObject countCmd = new BasicDBObject("count", "coll");
+  public void testUndefinedCommand() throws Exception {
     Fongo fongo = newFongo();
+    DB db = fongo.getDB("db");
+    CommandResult result = db.command("undefined");
+    assertEquals("ok should always be defined", 0.0, result.get("ok"));
+    assertEquals("no such cmd: undefined", result.get("errmsg"));
+  }
+
+  @Test
+  public void testCountCommand() throws Exception {
+    Fongo fongo = newFongo();
+
+    DBObject countCmd = new BasicDBObject("count", "coll");
     DB db = fongo.getDB("db");
     DBCollection coll = db.getCollection("coll");
     coll.insert(new BasicDBObject());
     coll.insert(new BasicDBObject());
     CommandResult result = db.command(countCmd);
-    assertEquals("The command should have been succesful", true, result.get("ok"));
-    assertEquals("The count should be in the result", 2L, result.get("n"));
+    assertEquals("The command should have been succesful", 1.0, result.get("ok"));
+    assertEquals("The count should be in the result", 2.0D, result.get("n"));
   }
 
   @Test
@@ -975,7 +1020,7 @@ public class FongoTest {
     DBObject countCmd = new BasicDBObject("count", "coll").append("limit", 2).append("skip", 4).append("query", builder.get());
     CommandResult result = db.command(countCmd);
     // Without sort.
-    assertEquals(0L, result.get("n"));
+    assertEquals(0D, result.get("n"));
   }
 
   @Test
@@ -1198,6 +1243,29 @@ public class FongoTest {
       // The _id field is always the first.
       assertEquals("_id", object.toMap().keySet().iterator().next());
     }
+  }
+
+  @Test
+  public void canInsertWithNewObjectId() throws Exception {
+    DBCollection collection = newCollection();
+    ObjectId id = ObjectId.get();
+
+    collection.insert(new BasicDBObject("_id", id).append("name", "jon"));
+
+    assertEquals(1, collection.count(new BasicDBObject("name", "jon")));
+    assertFalse(id.isNew());
+  }
+
+  @Test
+  public void saveStringAsObjectId() throws Exception {
+    DBCollection collection = newCollection();
+    String id = ObjectId.get().toString();
+
+    BasicDBObject object = new BasicDBObject("_id", id).append("name", "jon");
+    collection.insert(object);
+
+    assertEquals(1, collection.count(new BasicDBObject("name", "jon")));
+    assertEquals(id, object.get("_id"));
   }
 
   static class Seq {
